requirements.txt
google-api-python-client==1.6.2
google-cloud-datastore==1.0.0
google-cloud-storage==1.1.1
webapp2==2.5.2
jinja2==2.9.6
